requirements.txt
click==8.1.3
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pkg_resources==0.0.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
zipp==3.8.0
